ApiCoreService/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27
SQLAlchemy>=2.0,<2.1
asyncpg>=0.29
aio-pika>=9.4.0
python-dotenv==1.0.1
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
pydantic[email]>=2.6
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
